Invoke sendMessage() on the current event loop
A registered WebSocketService may invoke `ping(withMessage:)`,
`send(message:asBinary:)` and `send(message:)` from a thread different from the
EventLoop thread that drives the WebSocketConnection. These functions allocate
`ByteBuffer`s and write to the connected Channel. We must make sure that the
buffer allocator and `sendMessage()` are invoked on the EventLoop related to the
current WebSocketConnection.